move partition
