+++ ydb/core/persqueue/pqtablet/partition/mirrorer.cpp
@@ -1,5 +1,5 @@
 #include "mirrorer.h"
-#include "write_meta.h"
+#include <ydb/core/persqueue/write_meta.h>
 
 #include <ydb/core/persqueue/writer/source_id_encoding.h>
 #include <ydb/core/protos/grpc_pq_old.pb.h>
