+++ ydb/core/persqueue/pqtablet/partition/partition_init.cpp
@@ -1,4 +1,4 @@
-#include "offload_actor.h"
+#include <ydb/core/persqueue/offload_actor.h>
 #include "partition.h"
 #include "partition_compactification.h"
 #include "partition_log.h"
