+++ ydb/core/persqueue/pqtablet/partition/partition_read.cpp
@@ -1,11 +1,11 @@
-#include "event_helpers.h"
+#include <ydb/core/persqueue/event_helpers.h>
 #include "mirrorer.h"
 #include "partition_log.h"
 #include "partition_util.h"
 #include "partition_common.h"
 #include "partition_compactification.h"
-#include "read.h"
-#include "dread_cache_service/caching_service.h"
+#include <ydb/core/persqueue/read.h>
+#include <ydb/core/persqueue/dread_cache_service/caching_service.h>
 
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/blobstorage.h>
