+++ ydb/core/persqueue/pqtablet/transaction.cpp
@@ -1,6 +1,6 @@
 #include "transaction.h"
-#include "utils.h"
-#include "partition_log.h"
+#include <ydb/core/persqueue/utils.h>
+#include "pq_log.h"
 
 #include <ydb/library/wilson_ids/wilson.h>
