+++ ydb/core/persqueue/ut/sourceid_ut.cpp
@@ -1,4 +1,4 @@
-#include "sourceid.h"
+#include <ydb/core/persqueue/pqtablet/partition/sourceid.h>
 
 #include <ydb/core/keyvalue/keyvalue_events.h>
 #include <ydb/core/persqueue/key.h>
